final push in reworking field transformer / figure helper / metadata modules and methods
