+++ app/fig_meta/speaker_topic_scatter_meta.py
